requirements.txt
Django==1.6.2
South==0.8.4
dj-database-url==0.3.0
dj-static==0.0.5
django-bootstrap-form==3.1
djangorestframework==2.3.13
psycopg2==2.5.2
static==1.0.2
wsgiref==0.1.2
